fix(quality): enable suppressed checks, dedupe allowlist, add OCR
Widen ruff lint ruleset (E,W,F,I,N,B,BLE,UP,SIM,C4,RET,ARG,RUF) and fix
all 109 violations without suppressions. ARG per-file-ignored in tests
only: fixtures are requested for side effects (tmp_repo chdir) and mock
signatures mirror the interfaces they replace.

Delete the dead antigravity client (zero importers) and its fully
skipped test module; drop the dead antigravity/quota constructor params.

Extract EXTENDED_COGNATE_ALLOWLIST to extended_cognates.json as the
single source of truth; the dict lived as diverged copies in
check_data_contract.py (3492 words) and work/allowlist_extra.py
(3474 words), which pushed new-code duplication to 70.95% (gate: 3%).
work/clean_expansion.py now emits a JSON merge proposal.

Gate the SonarQube workflow on the self-hosted quality gate
(sonar.qualitygate.wait=true); the scan previously went green
regardless of SQ findings.

Port the AI Code Review (OpenCodeReview) workflow and scripts from
Vidiom; ENABLE_AI_CODE_REVIEW must be flipped on after merge because
the scripts run from the base branch copy (see Vidiom issue 2004).

+++ scripts/gemma_qa/progress.py
@@ -11,7 +11,7 @@ from collections.abc import Sequence
 def format_duration(seconds: float | None) -> str:
     if seconds is None or seconds < 0 or seconds != seconds:  # NaN
         return "?"
-    total = int(round(seconds))
+    total = round(seconds)
     hours, rem = divmod(total, 3600)
     minutes, secs = divmod(rem, 60)
     if hours:
